chore: Added removal of non-unique types in Json Schema type array
